requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
textual>=0.40.0
rich>=13.0.0
